luci-theme-openwrt-2020: remove max-width attribute
If the viewport width is higher than max-width value and lower than
screen width, tooltips for main pie charts in luci-app-nlbwmon
are shifted to the left. This bug does not occur with other themes
(at least luci-theme-bootstrap).
Removing this attribute fixes this and allows to expand view to the
full screen width (another option would be to set much higher value).